// Application/Frontend/frontend-app/src/app/generated/trade-client/model/balanceResponse.ts
/**
 * Trade REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BalanceResponse { 
    accountId: number;
    cashBalance: number;
    /**
     * ISO 4217 code.
     */
    currency: string;
    asOf: string;
}