// Application/Frontend/frontend-app/src/app/generated/trade-client/model/orderSide.ts
/**
 * Trade REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Direction of the order.
 */
export const OrderSide = {
    Buy: 'BUY',
    Sell: 'SELL',
} as const;
export type OrderSide = typeof OrderSide[keyof typeof OrderSide];